Enum debris_llir::class::ClassKind

pub enum ClassKind {
    Type(Type),
    Struct(StructRef),
    StructValue(StructRef),
    Tuple(TupleRef),
    TupleValue(TupleRef),
    Function(FunctionClassRef),
}

An enum over every type that exists in debris. Either a simple Type or a complex type with additional data.

Most complex types come in two variants: As a pattern and as a value. A pattern could be e.g. a struct which may be required in a function definition as parameter. The value would then be the struct object that can actually be passed to that function.

Variants

Type(Type)

Struct(StructRef)

StructValue(StructRef)

Tuple(TupleRef)

TupleValue(TupleRef)

Function(FunctionClassRef)

Implementations

impl ClassKind

pub fn as_value(&self) -> Option<ClassKind>

Changes all types to their respective objects, e.g. Struct => StructObject. This can be used to figure out the ClassKind that matches on this kind

Returns None if this is already an object variant

pub fn matches(&self, other: &ClassKind) -> bool

Returns whether the other class kind matches this class kind if this is interpreted as a pattern. For example, a struct object can match on a struct, if the underlying struct is equal.

pub fn diverges(&self) -> bool

Returns whether this value diverges, aka is impossible to construct. A diverging class can match on any pattern.

This is also used for some optimizations

pub fn typ(&self) -> Type

Returns the simple Type of this ClassKind

pub fn runtime_encodable(&self) -> bool

Returns whether this type can be fully encoded at runtime.

pub fn comptime_encodable(&self) -> bool

Returns whether this type can be fully encoded at compile time.