Module debris_llir::debris_object

Structs

DebrisObject

Objects are a central type for the compiler. Basically anything that can be assigned to a variable is an object. This includes numbers, function, modules, and more. It is possible to cast the ObjectPayload to its original value. Because DebrisObject is unsized, it generally only accessed as ObjectRef

ObjectRef

This struct is used to pass objects around

Traits

ObjectPayload

A trait for values that can be used as debris object payload

ValidPayload

Type Definitions

ObjectProperties

The type of the properties map