Struct debris_llir::llir_builder::LlirBuilder

pub struct LlirBuilder<'ctx> {
    pub(crate) compile_context: &'ctx CompileContext,
    pub(crate) type_context: TypeContext,
    pub(crate) functions: FxHashMap<BlockId, Function>,
    pub(crate) compiled_contexts: FxHashMap<MirContextId, BlockId>,
    pub(crate) native_functions: Vec<FunctionGenerics<'ctx>>,
    pub(crate) runtime: Runtime,
    pub(crate) block_id_generator: BlockIdGenerator,
    pub(crate) global_namespace: &'ctx MirNamespace,
    pub(crate) return_values_arena: &'ctx ReturnValuesArena,
    pub(crate) object_mapping: FxHashMap<MirObjectId, ObjectRef>,
    pub(crate) item_id_allocator: ItemIdAllocator,
    pub(crate) call_stack: CallStack,
}

Fields
compile_context: &'ctx CompileContexttype_context: TypeContextfunctions: FxHashMap<BlockId, Function>compiled_contexts: FxHashMap<MirContextId, BlockId>native_functions: Vec<FunctionGenerics<'ctx>>

A list of all used native functions and their instantiations

runtime: Runtimeblock_id_generator: BlockIdGeneratorglobal_namespace: &'ctx MirNamespacereturn_values_arena: &'ctx ReturnValuesArenaobject_mapping: FxHashMap<MirObjectId, ObjectRef>item_id_allocator: ItemIdAllocatorcall_stack: CallStack
Implementations

impl<'ctx> LlirBuilder<'ctx>

pub fn new(
ctx: &'ctx CompileContext,
extern_items_factory: impl Fn(&TypeContext) -> HashMap<Ident, ObjectRef>,
mir_extern_items: &FxHashMap<Ident, MirExternItem>,
namespace: &'ctx MirNamespace,
return_values_arena: &'ctx ReturnValuesArena
) -> Result<Self>

pub fn build(
self,
entry_context_id: MirContextId,
contexts: &'ctx FxHashMap<MirContextId, MirContext>
) -> Result<Llir>

pub(crate) fn get_obj(&self, obj_id: MirObjectId) -> ObjectRef

pub(crate) fn get_obj_opt(&self, obj_id: MirObjectId) -> Option<ObjectRef>

pub(crate) fn _set_obj(&mut self, obj_id: MirObjectId, value: ObjectRef)

pub fn compile_context(
&mut self,
contexts: &'ctx FxHashMap<MirContextId, MirContext>,
context_id: MirContextId
) -> Result<(BlockId, ObjectRef)>

pub fn force_compile_context(
&mut self,
contexts: &'ctx FxHashMap<MirContextId, MirContext>,
context_id: MirContextId,
block_id: BlockId
) -> Result<(BlockId, ObjectRef)>

Compiles any context, even if it is already compiled. Force compiling a context does not insert it into the list of compiled contexts This is used for example when monomorphizing functions, where the same function gets evaluated with different compile time values block_id is the id of the block which will get populated by this function