Enum debris_llir::llir_nodes::Condition

pub enum Condition {
    Compare {
        lhs: ScoreboardValue,
        rhs: ScoreboardValue,
        comparison: ScoreboardComparison,
    },
    And(Vec<Condition>),
    Or(Vec<Condition>),
}

Evaluates a condition and returns either true or false

Variants

Compare

Fields

lhs: ScoreboardValue

rhs: ScoreboardValue

comparison: ScoreboardComparison

Comparison between two values, eg. val1 <= val2

And(Vec<Condition>)

Or(Vec<Condition>)

Implementations

impl Condition

pub fn not(&self) -> Condition

Returns a condition that is true, when this condition is false and the other way around

pub fn is_effect_free(&self) -> bool

Returns whether evaluating this condition has any side effects

pub fn is_simple(&self) -> bool

Checks whether this condition is “simple”, which means that it is not expensive to use this condition multiple times instead of a boolean.

Right now, only Condition::Compare is considered simple.

pub fn variable_reads(&self) -> FxHashSet<ItemId>

pub fn variable_accesses<F: FnMut(VariableAccess<'_>)>(&self, visitor: &mut F)

Recursively yields all variables that this condition reads from

pub fn variable_accesses_mut<F: FnMut(VariableAccessMut<'_>)>(
&mut self,
visitor: &mut F
)