Module debris_llir::minecraft_utils

Enums

Scoreboard

Identifies a specific scoreboard

ScoreboardComparison

Any comparison that can be executed on two scoreboard values

ScoreboardOperation

Any operation that can be executed on a scoreboard- This excludes copy, because scoreboard operations are assumed to be trinary (a = b OP c), but copy is only a binary Operation.

ScoreboardValue

A Value that can be stored on a scoreboard