Struct debris_llir::objects::obj_module::ObjModule

pub struct ObjModule {
    pub ident: Ident,
    pub members: ObjectProperties,
}

A module object

Contains other values, including nested modules.

Fields
ident: Ident

The identifying name of this module

members: ObjectProperties

The members of this module

Implementations

impl ObjModule

pub fn new(name: impl Into<Ident>) -> Self

Creates a new empty module with this name

pub fn with_members(name: Ident, members: ObjectProperties) -> Self

pub fn ident(&self) -> &Ident

Returns the ident of this module

pub fn members(&self) -> impl Iterator<Item = (&Ident, &ObjectRef)>

pub fn set_property<T: Into<Ident>>(
&mut self,
name: T,
value: ObjectRef
) -> Option<ObjectRef>

Sets a property

If it already exists, replaces the old value and returns it

pub fn register<T: Into<Ident>>(&mut self, name: T, value: ObjectRef)

Registers a value for the first time

Panics if it already exists

pub fn register_function(&mut self, ctx: &TypeContext, function: ObjFunction)

A more concise way to register builtin functions without having to declare its name twice.