Struct debris_llir::opt::global_opt::Commands

pub struct Commands<'opt, 'ctx> {
    pub optimizer: &'opt mut GlobalOptimizer<'ctx>,
    pub stats: &'opt mut CodeStats,
    pub infinite_loop_detector: InfiniteLoopDetector,
    pub commands: &'opt mut OptimizeCommandDeque<OptimizeCommand>,
}

Interface for optimizing functions to get data about the code and emit optimization instructions

Fields
optimizer: &'opt mut GlobalOptimizer<'ctx>stats: &'opt mut CodeStatsinfinite_loop_detector: InfiniteLoopDetectorcommands: &'opt mut OptimizeCommandDeque<OptimizeCommand>
Implementations

impl<'opt, 'ctx> Commands<'opt, 'ctx>

fn new( optimizer: &'opt mut GlobalOptimizer<'ctx>, stats: &'opt mut CodeStats, commands: &'opt mut OptimizeCommandDeque<OptimizeCommand> ) -> Self

fn retain_functions(&mut self)

pub fn is_id_unused(&self, id: ItemId, node: (BlockId, usize)) -> bool

pub fn get_info(&self, var: ItemId) -> &VariableUsage

Returns the variable info for this node

pub fn get_reads(&self, var: ItemId) -> usize

pub fn get_call_count(&self, function: BlockId) -> usize

fn run_optimizer<F>(&mut self, optimizer: &mut F) -> boolwhere F: Optimizer,

Runs an optimizing function

Returns whether this function could optimize anything

fn execute_commands(&mut self)

Execute every command that is in the current command stack