Module debris_llir::opt::optimizers::redundancy_optimizer

Structs

RedundancyOptimizer

Removes useless nodes

Enums

SimplifiedCondition

Functions

merge_condition

Tries to merge the condition and the previous node into one node eg: a := b > c; d := a == 1 => d := b > c

simplify_condition

Simplifies trivial conditions

write_after_write

Checks for a write to id after node and returns with false If it is read before a write. Also returns false if no further write in this branch exists