Struct debris_llir::opt::variable_metadata::ValueHints

pub struct ValueHints {
    hints: FxHashMap<ItemId, Hint>,
}

Stores the possible range of values of a runtime variable

Fields
hints: FxHashMap<ItemId, Hint>
Implementations

impl ValueHints

pub fn set_hint(&mut self, id: ItemId, hint: Hint)

pub fn clear_hint(&mut self, id: ItemId)

Clears all hints for this id

pub fn clear_all(&mut self)

Clears all hints

pub fn get_hint(&self, id: ItemId) -> Hint

pub fn get_scoreboard_value(&self, value: &ScoreboardValue) -> Option<i32>

Tries to get the static value of a scoreboard value

pub fn update_hints(&mut self, node: &Node, guaranteed_run: bool)

Updates the hints for all variables that this node modifies guaranteed_run specifies whether node will definitely execute

pub fn static_binary_operation(&self, bin_op: &BinaryOperation) -> Option<i32>

Tries to evaluate a binary operation with static values

pub fn static_condition(&self, condition: &Condition) -> Option<bool>

Tries to evaluate a condition with static values