Crate debris_llir

Low-level intermediate representation

The llir is quite close to minecraft functions.

Modules

block_id

class

This module defines classes. The concept of a class should be similar to the equivalent concept in other languages. Every value (object) has a class. Builtin values have a class with an associated type, while user defined classes, like structs, carry a reference to the concrete struct.

debris_object 🔒

error_utils 🔒

extern_item_path

function_interface

This module contains a trait ToFunctionInterface which allows to write nice function interfaces by using much less boilerplate to define functions.

item_id

json_format

(Current incomplete) interface for minecraft’s text commands.

llir_builder 🔒

llir_function_builder 🔒

llir_impl 🔒

llir_nodes

Contains every node that can be produced in the llir step.

llir_shared_state 🔒

macro_impl_class 🔒

memory 🔒

minecraft_utils

objects

Contains all objects which can be used in a debris program.

opt 🔒

Contains two optimizer that can operate on the llir: The peephole optimizer which optimizes every emitted node on its own (but is also able to look at previous nodes) and the global optimizer

runtime 🔒

type_context

types 🔒

Macros

impl_class

log

match_object

Helper macro that makes match-style semantics available for duck typed objects

Structs

CallGraph

CodeStats

tracks statistics about the global code which can be used to allow some optimizations

DebrisObject

Objects are a central type for the compiler. Basically anything that can be assigned to a variable is an object. This includes numbers, function, modules, and more. It is possible to cast the ObjectPayload to its original value. Because DebrisObject is unsized, it generally only accessed as ObjectRef

Llir

The low-level intermediate representation struct

ObjectRef

This struct is used to pass objects around

Runtime

Enums

Type

The type of a class object

Traits

ObjectPayload

A trait for values that can be used as debris object payload

ValidPayload

Type Definitions

NativeFunctionId

ObjectProperties

The type of the properties map