Trait debris_llir::debris_object::ObjectPayload

pub trait ObjectPayload: ValidPayload {
    fn memory_layout(&self) -> &MemoryLayout;

    fn create_class(&self, ctx: &TypeContext) -> ClassRef { ... }

    fn runtime_class(&self, _ctx: &TypeContext) -> Option<ClassRef> { ... }

    fn get_property(
        &self, 
        _ctx: &TypeContext, 
        _ident: &Ident
    ) -> Option<ObjectRef> { ... }
}

A trait for values that can be used as debris object payload

The private ValidPayload trait is auto-implemented

Required methods

fn memory_layout(&self) -> &MemoryLayout

Returns the memory layout of this specific object This method is usually only called once

Provided methods

fn create_class(&self, ctx: &TypeContext) -> ClassRef

The class specific to this object. Contains additionally to the class generics and the memory layout

Per default the default class of the object type

fn runtime_class(&self, _ctx: &TypeContext) -> Option<ClassRef>

Returns the class which the runtime encodable variant of this object would have. Used to determine the class passed into the promote method of objects.

fn get_property(&self, _ctx: &TypeContext, _ident: &Ident) -> Option<ObjectRef>

May be overwritten by distinct payloads which carry properties