Struct debris_llir::llir_builder::LlirBuilder

pub struct LlirBuilder<'ctx> {
    extern_items: FxHashMap<MirObjectId, ObjectRef>,
    pub(crate) compile_context: &'ctx CompileContext,
    pub(crate) type_context: TypeContext,
    pub(crate) native_function_map: NativeFunctionMap<'ctx>,
    pub(crate) shared_states: SharedStates,
    pub(crate) runtime: Runtime,
    pub(crate) block_id_generator: BlockIdGenerator,
    pub(crate) global_namespace: &'ctx MirNamespace,
    pub(crate) return_values_arena: &'ctx ReturnValuesArena,
    pub(crate) item_id_allocator: ItemIdAllocator,
}

Fields
extern_items: FxHashMap<MirObjectId, ObjectRef>compile_context: &'ctx CompileContexttype_context: TypeContextnative_function_map: NativeFunctionMap<'ctx>

Stores the already compiled functions This is not part of the llir function builder shared state, because a computed function should always be the same, no matter where it is computed from

shared_states: SharedStates

A map of shared state id to shared state

runtime: Runtimeblock_id_generator: BlockIdGeneratorglobal_namespace: &'ctx MirNamespacereturn_values_arena: &'ctx ReturnValuesArenaitem_id_allocator: ItemIdAllocator
Implementations

impl<'ctx> LlirBuilder<'ctx>

pub fn new(
ctx: &'ctx CompileContext,
extern_items_factory: impl Fn(&TypeContext) -> HashMap<Ident, ObjectRef>,
mir_extern_items: &FxHashMap<Ident, MirExternItem>,
namespace: &'ctx MirNamespace,
return_values_arena: &'ctx ReturnValuesArena
) -> Result<Self>

pub fn build(
self,
entry_context_id: MirContextId,
contexts: &'ctx FxHashMap<MirContextId, MirContext>
) -> Result<Llir>