Module debris_llir::opt

Contains two optimizer that can operate on the llir: The peephole optimizer which optimizes every emitted node on its own (but is also able to look at previous nodes) and the global optimizer

Modules

call_graph

code_stats

function_parameters

global_opt

logger

Simple logger to make debugging the optimizer simpler. Can be enabled using a flag.

optimize_commands

optimizers

This module contains various types implementing the Optimizer trait.

peephole_opt

variable_metadata

This module contains structs which can hold metadata for variables. This includes so called ValueHints which allow to deduce the set of possible runtime values a variable can have. This module also contains variable usage data which contain information about the amount of reads and writes a specific variable gets

Type Definitions

NodeId

The optimizer can uniquely identify each node with this type