Enum debris_llir::opt::optimize_commands::OptimizeCommandKind

pub enum OptimizeCommandKind {
    Delete,
    DiscardResult,
    InlineBranch(bool),
    UpdateBranch {
        branch: bool,
        new_node: Node,
    },
    UpdateBranchCondition(Condition),
    InlineFunction,
    RemoveAliasFunction(BlockId),
    ChangeWrite(ItemId),
    ChangeReads(ItemId, ScoreboardValue),
    SetCondition(Condition, Vec<usize>),
    Replace(Node),
    InsertAfter(Node),
}

Optimizing functions output commands that tell the optimizer what to do, this is done so that there are no troubles with mutability

Variants

Delete

Deletes a single node

DiscardResult

Converts the FastStoreFromResult node into its command, discarding the result

InlineBranch(bool)

Tuple Fields

0: bool

Discards the node and only keeps the branch that matches the bool. (true => pos_branch, false => neg_branch)

UpdateBranch

Fields

branch: bool

new_node: Node

Updates the specified branch with the new node

UpdateBranchCondition(Condition)

Tuple Fields

0: Condition

Updates the condition of a branch

InlineFunction

Inlines the function of this function call.

RemoveAliasFunction(BlockId)

Tuple Fields

0: BlockId

Removes all aliases to a function which only redirects to another function The argument specifies the aliased function.

ChangeWrite(ItemId)

Tuple Fields

0: ItemId

Changes the variable this node writes to

ChangeReads(ItemId, ScoreboardValue)

Tuple Fields

0: ItemId

1: ScoreboardValue

Replaces all variables .0 with .1

SetCondition(Condition, Vec<usize>)

Tuple Fields

0: Condition

1: Vec<usize>

Changes the condition of this branch to a new condition Vec usize contains the exact index of the condition to replace (Vec since conditions can be nested)

Replace(Node)

Tuple Fields

0: Node

Replaces the old node completely

InsertAfter(Node)

Tuple Fields

0: Node

Inserts this node after