Module debris_llir::opt::optimizers::redundant_copy_optimizer

TODO: implement this properly

Structs

RedundantCopyOptimizer

A node which copies a value a to b is often redundant. This is especially important, because the compiler tends to generate operations in the shape of let temp = a; operation_with_temp(temp); a = temp. This optimizer aggressively removes copy instructions (let temp = a) and changes all subsequent reads to the previous variable (a).