Struct debris_llir::opt::peephole_opt::PeepholeOptimizer

pub struct PeepholeOptimizer {
    nodes: Vec<Node>,
    opt_mode: OptMode,
    value_hints: ValueHints,
}

A just-in-time peephole optimizer.

This means, that the optimizer optimizes nodes as they are being added, while also having access to the previously emitted nodes.

A single PeepholeOptimizer can only be active in a single context (aka a single .mcfunction file)

Fields
nodes: Vec<Node>opt_mode: OptModevalue_hints: ValueHints

Information about the possible values of runtime variables

Implementations

impl PeepholeOptimizer

pub fn from_compile_context(ctx: &CompileContext) -> Self

pub fn push(&mut self, node: Node)

Adds this node to the collection and optimizes it on the fly

pub fn _nodes(&self) -> &[Node]

pub fn take(&mut self) -> Vec<Node>

Drops this instance and returns the wrapped nodes

impl PeepholeOptimizer

fn optimize_and_insert(&mut self, node: Node)

Optimizes a node and the previous nodes and pushes the new node to the collection of nodes

This operation may affect already pushed nodes.

fn optimize(&mut self, node: Node) -> Node

fn optimize_branch(&mut self, branch: Branch) -> Node